+++ src/utils.js
@@ -1,3 +1,5 @@
+/// <reference path='./types.js' />
+
 /**
  * Waits for a given amount of time
  * @param {number} ms The time in milliseconds to wait before resuming execution
@@ -61,4 +63,26 @@ export const getElements = parentElement => {
 			.filter(el => hasSingleTextNode(el))
 			.map(currentNode => ({ currentNode, text: currentNode.textContent.split('') }))
 	}
+}
+
+/** @type {TypewriterEffectFn} */
+export const loopTypewriterEffect = async ({ currentNode, text }, options) => {
+	currentNode.textContent = ''
+	currentNode.classList.add('typing')
+	for (const letter of text) {
+		currentNode.textContent += letter
+		const fullyWritten = currentNode.textContent === text.join('')
+		if (fullyWritten) {
+			options.dispatch('done')
+			await sleep(typeof options.loop === 'number' ? options.loop : 1500)
+			while (currentNode.textContent !== '') {
+				currentNode.textContent = currentNode.textContent.slice(0, -1)
+				await typingInterval(options.interval)
+			}
+		}
+		await typingInterval(options.interval)
+	}
+	currentNode.nextSibling !== null && currentNode.classList.length == 1
+		? currentNode.removeAttribute('class')
+		: currentNode.classList.remove('typing')
 }